refactor(#360): extract shared QC role/validation helpers to tools/_qc_shared.py
Move _role_kwargs + _validate_trait_subset out of qc_clean_tool.py into tools/_qc_shared.py so qc_clean and the incoming qc_inspect forward detected role columns and reject bad trait_columns identically. Pure move — qc_clean behavior and its suite are unchanged.
