+++ bloommcp/src/bloom_mcp/tools/qc_clean_tool.py
@@ -40,16 +40,16 @@ from __future__ import annotations
 import json
 from typing import Optional
 
-import pandas as pd
 from pydantic import BaseModel, Field
 from sleap_roots_analyze import clean_traits_for_analysis
 
 from bloom_mcp.contract import BloomMCPError, Provenance, as_mcp_tool
 from bloom_mcp.contract import register as _contract_register
-from bloom_mcp.data_access import ExperimentFrame, ExperimentReadError
+from bloom_mcp.data_access import ExperimentReadError
 from bloom_mcp.data_utils import convert_to_json_serializable
 from bloom_mcp.experiment_utils import CLEANED_CSV_NAME, TRAITS_DIR
 from bloom_mcp.tools import _ports
+from bloom_mcp.tools._qc_shared import _role_kwargs, _validate_trait_subset
 
 _TOOL_CLASS = "qc"
 _LOG_NAME = "cleanup_log.json"
@@ -147,47 +147,6 @@ class QCCleanResult(BaseModel):
     version_dir: str
     manifest_path: str
     outputs: dict[str, str]
-
-
-def _role_kwargs(frame: ExperimentFrame) -> dict[str, str]:
-    """Forward the adapter-detected role columns to the delegate.
-
-    Omit any role that is ``None`` so ``clean_traits_for_analysis`` applies its
-    own default rather than receiving ``None``.
-    """
-    roles = {
-        "barcode_col": frame.sample_id_col,
-        "genotype_col": frame.genotype_col,
-        "replicate_col": frame.replicate_col,
-    }
-    return {k: v for k, v in roles.items() if v is not None}
-
-
-def _validate_trait_subset(
-    frame: ExperimentFrame, requested: list[str], experiment: str
-) -> None:
-    """Reject a caller-supplied ``trait_columns`` subset up front with a clear remedy.
-
-    Without this an unknown column raises ``KeyError`` (→ opaque ``internal_error``)
-    and a non-numeric column silently corrupts the delegate's zero/NaN-fraction
-    filtering. Both surface here as a fixable ``invalid_input`` naming the columns.
-    """
-    missing = [c for c in requested if c not in frame.df.columns]
-    if missing:
-        raise BloomMCPError(
-            code="invalid_input",
-            message=f"trait_columns names columns not in {experiment!r}: {missing}.",
-            remedy="Use column names from load_experiment_data, or omit trait_columns to clean all detected traits.",
-        )
-    non_numeric = [
-        c for c in requested if not pd.api.types.is_numeric_dtype(frame.df[c])
-    ]
-    if non_numeric:
-        raise BloomMCPError(
-            code="invalid_input",
-            message=f"trait_columns includes non-numeric columns: {non_numeric}.",
-            remedy="Pass only numeric trait columns; metadata/identifier columns cannot be cleaned as traits.",
-        )
 
 
 @as_mcp_tool(
